Update #37 Nearing Designer Block finish Line
Designer blocks need an associated Mesh, AssetLoader needs a way to associate InstanceGridTextures, and the render code needs to be cleaned and ridded of the nulls I added just to make it work.

Future work towards making a world.renderScene() needs to be made, and uniform, since the Client and Server should use the same rendering stuff for blocks and entities.

Transparency still needs to be added :_:

+++ src/com/GlitchyDev/Rendering/Assets/WorldElements/GameItem.java
@@ -2,7 +2,6 @@ package com.GlitchyDev.Rendering.Assets.WorldElements;
 
 import com.GlitchyDev.Rendering.Assets.Mesh.Mesh;
 import org.joml.Vector3f;
-import org.joml.Vector3i;
 
 /**
  * A object that can be represented in game in 3D or 2D ( See SpriteItem or HudItem )
@@ -37,7 +36,7 @@ public class GameItem {
         this.position.z = z;
     }
 
-    public void setPosition(Vector3i newPosition) {
+    public void setPosition(Vector3f newPosition) {
         this.position.set(newPosition);
     }
